requirements.txt
flask==3.0.0
requests==2.31.0
yfinance==0.2.32
pandas==2.1.3
numpy==1.26.2
ta==0.11.0
python-dateutil==2.8.2
ollama==0.1.7
python-dotenv